requirements.txt
Django==1.7
Markdown==2.5.1
Pygments==1.6
argparse==1.2.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.8
django-toolbelt==0.0.1
djangorestframework==2.4.3
gunicorn==19.1.1
mimeparse==0.1.3
nltk==3.0.0
numpy==1.9.1
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.4.3
six==1.8.0
static3==0.5.1
textblob==0.9.0
wsgiref==0.1.2
